pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maho"
version = "0.1.0"
description = "Maho - A refactored Agent Zero fork with proper architecture"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Maho Contributors" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    # Web framework and ASGI
    "a2wsgi==1.10.8",
    "flask[async]==3.0.3",
    "flask-basicauth==0.2.0",
    # AI/ML frameworks (CPU-only, no CUDA)
    "langchain-anthropic==0.3.3",
    "langchain-community==0.3.19",
    "langchain-google-genai==2.1.2",
    "langchain-groq==0.2.2",
    "langchain-huggingface==0.1.2",
    "langchain-mistralai==0.2.4",
    "langchain-ollama==0.3.0",
    "langchain-openai==0.3.11",
    "tiktoken==0.8.0",
    "sentence-transformers==3.0.1",
    # Speech-to-text
    "openai-whisper==20240930",
    # Vector database and search
    "faiss-cpu==1.11.0",
    "duckduckgo-search==6.1.12",
    # Browser automation and web scraping
    "browser-use==0.2.5",
    "playwright==1.52.0",
    "newspaper3k==0.2.8",
    "lxml_html_clean==0.3.1",
    # Document processing
    "pypdf==4.3.1",
    "unstructured[all-docs]==0.16.23",
    "unstructured-client==0.31.0",
    "langchain-unstructured[all-docs]==0.1.6",
    "markdown==3.7",
    "markdownify==1.1.0",
    "pymupdf==1.25.3",
    "pytesseract==0.3.13",
    "pdf2image==1.17.0",
    # Infrastructure and DevOps
    "docker==7.1.0",
    "paramiko==3.5.0",
    "GitPython==3.1.43",
    # MCP (Model Context Protocol)
    "mcp==1.9.0",
    "fastmcp==2.3.4",
    "flaredantic==0.1.4",
    # Async framework
    "anyio[trio]>=4.0.0",
    "trio>=0.23.0",
    # Utilities
    "python-dotenv==1.1.0",
    "pytz==2024.2",
    "webcolors==24.6.0",
    "nest-asyncio==1.6.0", # Keep for compatibility during migration
    "crontab==1.0.1",
    "ansio==0.0.1",
    "inputimeout==1.0.4",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
    "jsonschema>=4.24.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]



[project.scripts]
maho-ui = "scripts.start_ui:main"
maho-tunnel = "scripts.start_tunnel:main"
maho-prepare = "scripts.prepare:main"
maho-preload = "scripts.preload:main"

[project.urls]
Homepage = "https://github.com/omniwaifu/maho"
Repository = "https://github.com/omniwaifu/maho"
Issues = "https://github.com/omniwaifu/maho/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501",  # line too long, handled by black
    "E402",  # module level import not at top of file
    "F401",  # imported but unused
    "F841",  # local variable assigned but never used
    "I001",  # import block is un-sorted or un-formatted
    "W293",  # blank line contains whitespace
    "W291",  # trailing whitespace
    "W292",  # no newline at end of file
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ["py310"]
line-length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "browser_use.*",
    "newspaper.*",
    "unstructured.*",
    "faiss.*",
    "whisper.*",
    "ansio.*",
]
ignore_missing_imports = true 
